김민석 : 벨리알 투낙 스포너 및 시작이벤트 추가

+++ GameEngineContents/GroundBomb.h
@@ -33,7 +33,7 @@ private:
 	float4 BGScale = { 150.0f,800.0f,0.0f };
 	float4 BombScale = { 300.0f,800.0f,0.0f };
 	float4 BombColScale = { 200.0f,800.0f,0.0f };
-	float GroundBombDamage = 5;
+	int GroundBombDamage = 5;
 	float Time = 0.0f;
 	float StartAlpha = 0.6f;
 	float AlphaTime = 0.0f;
